requirements.txt
tqdm==4.30.0
scipy==1.2.0
matplotlib==3.0.3
numpy==1.16.2
Pillow==5.4.1
scikit-learn==0.20.3
